chore(medusa): Update request-ip dependency version (#4636)
This updates the request-ip package dependency in @medusajs/medusa.

This resolves the critical security warnings seen when creating a new medusa backend OTHER THAN medusa-admin related issues.  Because the admin is undergoing an extensive rewrite, I didn't address the older version of node-fetch that is at the root of those issues.

Co-authored-by: Oli Juhl <59018053+olivermrbl@users.noreply.github.com>