fix: eslint configurations in the root (#6049)
When running `yarn lint` in the root of the monorepo, we get the following error:

```bash
Oops! Something went wrong! :(

ESLint: 8.44.0

Error: ESLint configuration in .eslintrc.js is invalid:
	- Unexpected top-level property "overrides[4].ignorePatterns".
```

Rmoving the `ignorePatterns` in configurations for `packages/admin-next/dashboard/**/*` and instead adding the path `packages/admin-next/dashboard/**/dist` fixes the issue.