feat(admin-sdk): Merge users Vite config (#7509)
**What**
- Allows users to customize the Vite config from `admin.vite` in `medusa-config.js`

**How**
```
/** @type {import('@medusajs/types').ConfigModule} */
module.exports = {
  projectConfig,
  admin: {
    backendUrl: "http://localhost:9000",
    // the config param is our default config
    vite: (config) => {
      return {
        define: {
          __TEST__: JSON.stringify("test"), // this will now be defined as a const that can be accessed in widgets
        },
      };
    },
  },
 // ...
};
```

The `vite` param is a callback that takes our default config as a param. If the user does not need to access our config they don't have to use it, as we merge their config with our own, but accessing the default config can be useful in some cases.